FUlly working texture projection
Former-commit-id: 12b72ebd7d79f49a0968f96685f553a28f9cddad [formerly f7b4d8e543fe14b49f2fbc0c9e5be27b6478aaff]
Former-commit-id: 204722ea3efcab0f3e3ba64f41189c2e899c6259
Former-commit-id: 325cf2da890e909f636fde1cafa7b280a8b18e79
Former-commit-id: 3db505a378e5bcc6aee6c81b6b39af2ced93266f

+++ src/engine/spatial/mesh.cpp
@@ -68,7 +68,7 @@ void Mesh::change_basis_multithreaded(const std::list<Mesh*> mesh_list,
 
   for (const auto& mesh : mesh_list) {
     unsigned size = mesh->global_coordenates_faces.size();
-    unsigned segment = size / N_THREADS;
+    double segment = double(size) / N_THREADS;
 
     auto& m = MultithreadManager::get_instance();
     m.calculate_threaded(N_THREADS, [&](unsigned i) {
